refactored MediaPlayerRibbonComponent to MediaPlayer
