+++ src/components/media-player/media-player.component.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import classNames from 'classnames/bind';
 import { Col, Container, Row } from 'react-bootstrap';
 
-import styles from './media-player-ribbon.component.css';
+import styles from './media-player.component.css';
 import { MediaPlayerInfo } from './media-player-info.component';
 import { MediaPlayerControls } from './media-player-controls.component';
 import { MediaPlayerProgress } from './media-player-progress.component';
@@ -10,7 +10,7 @@ import { MediaPlayerSide } from './media-player-side.component';
 
 const cx = classNames.bind(styles);
 
-export function MediaPlayerRibbonComponent() {
+export function MediaPlayer() {
   return (
     <Container fluid className={cx('h-100')}>
       <Row className={cx('media-player-container')}>
